Refactor getLocationReplacements just for better code organization
