+++ src/ManageOptions.php
@@ -7,6 +7,7 @@ namespace Grasmash\ComposerScaffold;
 use Composer\Composer;
 use Composer\IO\IOInterface;
 use Composer\Package\PackageInterface;
+use Composer\Util\Filesystem;
 use Grasmash\ComposerScaffold\Operations\OperationInterface;
 use Grasmash\ComposerScaffold\ScaffoldFilePath;
 
@@ -54,6 +55,40 @@ class ManageOptions {
    */
   public function packageOptions(PackageInterface $package) : ScaffoldOptions {
     return ScaffoldOptions::create($package->getExtra());
+  }
+
+  /**
+   * GetLocationReplacements creates an interpolator for the 'locations' element.
+   *
+   * The interpolator returned will replace a path string with the tokens
+   * defined in the 'locations' element.
+   *
+   * Note that only the root package may define locations.
+   *
+   * @return Interpolator
+   *   Object that will do replacements in a string using tokens in 'locations' element.
+   */
+  public function getLocationReplacements() : Interpolator {
+    return (new Interpolator())->setData($this->ensureLocations());
+  }
+
+  /**
+   * Ensure that all of the locatons defined in the scaffold filed exist.
+   *
+   * Create them on the filesystem if they do not.
+   */
+  public function ensureLocations() : array {
+    $fs = new Filesystem();
+    $locations = $this->getOptions()->locations() + ['web_root' => './'];
+    $locations = array_map(
+      function ($location) use ($fs) {
+        $fs->ensureDirectoryExists($location);
+        $location = realpath($location);
+        return $location;
+      },
+      $locations
+    );
+    return $locations;
   }
 
 }
