+++ src/ScaffoldOptions.php
@@ -5,7 +5,6 @@ declare(strict_types = 1);
 namespace Grasmash\ComposerScaffold;
 
 use Composer\IO\IOInterface;
-use Composer\Util\Filesystem;
 use Grasmash\ComposerScaffold\Operations\OperationInterface;
 use Grasmash\ComposerScaffold\ScaffoldFilePath;
 
@@ -168,25 +167,6 @@ class ScaffoldOptions {
       throw new \Exception($message);
     }
     return $this->getLocation($name);
-  }
-
-  /**
-   * Ensure that all of the locatons defined in the scaffold filed exist.
-   *
-   * Create them on the filesystem if they do not.
-   */
-  public function ensureLocations() : array {
-    $fs = new Filesystem();
-    $locations = $this->locations() + ['web_root' => './'];
-    $locations = array_map(
-      function ($location) use ($fs) {
-        $fs->ensureDirectoryExists($location);
-        $location = realpath($location);
-        return $location;
-      },
-      $locations
-    );
-    return $locations;
   }
 
   /**
